pyproject.toml
[project]
name = "legal-eagle-ai"
version = "0.1.0"
description = "Legal Eagle AI Project"
authors = [
    {name = "praveena0506", email = "praveenasubramaniyam0506@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi==0.117.1",        # Pinned to match Airflow requirements
    "uvicorn>=0.34.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "apache-airflow>=2.10.0",  # Using stable 2.10 branch
    "chromadb (>=1.3.7,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "datasets (>=4.4.2,<5.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false